--- src/pages/Login/NonMembersLogin.js
@@ -1,48 +1,0 @@
-import { Component } from "react";
-import "./Login.scss";
-import "./NonMembersLogin.scss";
-import { FaUserCircle } from "react-icons/fa";
-import { MdLock } from "react-icons/md";
-
-class NonMembersLogin extends Component {
-  render() {
-    return (
-      <div className="NonMembersLogin">
-        <div className="loginWrap">
-          <form className="loginForm">
-            <div className="LoginInput">
-              <form className="loginForm">
-                <span className="loginIconContainer">
-                  <FaUserCircle size="24" className="userIcon" />
-                  <input
-                    className="idInput"
-                    type="text"
-                    placeholder="주문자명"
-                  />
-                </span>
-                <div className="inputBoxSapce">
-                  <span className="loginIconContainer">
-                    <MdLock size="24" className="lockIcon" />
-                    <input
-                      className="pwInput"
-                      type="password"
-                      placeholder="주문번호"
-                    />
-                  </span>
-                </div>
-              </form>
-            </div>
-            <div className="loginButtonWrap">
-              <button className="loginButton">로그인</button>
-            </div>
-            <p className="caution">
-              주문번호와 비밀번호를 잊으신 경우, 고객센터로 문의하여 주시기
-              바랍니다.
-            </p>
-          </form>
-        </div>
-      </div>
-    );
-  }
-}
-export default NonMembersLogin;
